requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.2
frozenlist==1.4.1
google-ai-generativelanguage==0.1.0
google-api-core==2.21.0
google-auth==2.35.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.135
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.51.2
orjson==3.10.7
packaging==24.1
prompt-toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.10.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.31.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.2
